healthd: Add support for HVDCP, HVDCP_3, and Wipower chargers
HVDCP charger is high voltage DCP chargers. Add support to recognize
it and treat them as AC power source.

HVDCP_3 is a high voltage DCP charger where the charger's voltage
can be changed by issuing pulses on the D+/D- lines. Add support to
recognize it and treat it as an AC power source.

Wipower charging is a wireless charger where the power transfer
is via resonance and power control messages are exchanged over BLE.
Treat Wipower as a wireless charger.

Change-Id: Ib719529904e8b7a676bbdc5f5953f0f9da6df3fa